pyproject.toml
[project]
name = "ujin-xac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt==3.2.0",
    "fastapi[all]>=0.115.12",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-dateutil>=2.9.0.post0",
    "python-jose>=3.4.0",
]
